requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0

# OpenAI and Hugging Face
openai>=1.12.0
datasets>=2.14.0
huggingface-hub>=0.20.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
dayplot>=0.3.0

# Progress bars
tqdm>=4.65.0

# Date handling
python-dateutil>=2.8.2
